Simplify handlers by removing `Maybe` return value
In the ongoing idea to simplify things in Halogen 6, I've been thinking
a lot about doing this lately. I don't think I've ever used a `Nothing`
return in an event handler based on the value of the `Event`.

I have very occasionally had a listener that returned `Nothing` based on
the current component state, but there are two ways that can be handled
even after removing this: either omitting the listener in the prop
construction of the HTML (with `guard` or something), or just checking
the component state when handling the raised action instead.

Similarly the handler for `slot`ted child components - I do occasionally
use `const Nothing` as the handler when setting things up, but I've
sometimes been bitten by that by forgetting to go back and handle the
messages. If the component emits no message then `absurd` still works as
a handler in the absence of the `Maybe`.

So basically: I think this is more ergonomic for the vast majority of
use cases for handlers. For the rarer occurrences where filtering was
use, it's still possible to do that, just there's an extra step to deal
with.

I did also consider adding primed variants of everything, that still had
`Event -> Maybe i`, but despite my love of primed things, I figured
leaving them out entirely would probably be nicer from a new user
perspective :wink: